+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSsprite.h
@@ -23,19 +23,19 @@
 namespace enigma_user
 {
 
-void draw_sprite(int spr,int subimg,double x,double y);
-void draw_sprite_stretched(int spr,int subimg,double x,double y,double w,double h);
+void draw_sprite(int spr, int subimg, float x, float y);
+void draw_sprite_stretched(int spr, int subimg, float x, float y, float w, float h);
 //enigma::4args int draw_sprite_tiled(ARG sprite,ARG2 subimg,ARG3 x,ARG4 y)
-void draw_sprite_part(int spr,int subimg,double left,double top,double width,double height,double x,double y);
-void draw_sprite_part_offset(int spr,int subimg,double left,double top,double width,double height,double x,double y);
-void draw_sprite_ext(int spr,int subimg,double x,double y,double xscale,double yscale,double rot,int color,double alpha);
+void draw_sprite_part(int spr, int subimg, float left, float top, float width, float height, float x, float y);
+void draw_sprite_part_offset(int spr, int subimg, float left, float top, float width, float height, float x, float y);
+void draw_sprite_ext(int spr, int subimg, float x, float y, float xscale, float yscale, double rot, int color, double alpha);
 //draw_sprite_tiled_ext(sprite,subimg,x,y,xscale,yscale,color,alpha)
-void draw_sprite_stretched_ext(int spr,int subimg,double x,double y,double w,double h, int blend, double alpha);
-void draw_sprite_part_ext(int spr,int subimg,double left,double top,double width,double height,double x,double y,double xscale,double yscale,int color,double alpha);
-void draw_sprite_general(int spr,int subimg,double left,double top,double width,double height,double x,double y,double xscale,double yscale,double rot,int c1,int c2,int c3,int c4,double a1, double a2, double a3, double a4);
+void draw_sprite_stretched_ext(int spr, int subimg, float x, float y, float w, float h, int blend, double alpha);
+void draw_sprite_part_ext(int spr, int subimg, float left, float top, float width, float height, float x, float y, float xscale, float yscale, int color, double alpha);
+void draw_sprite_general(int spr, int subimg, float left, float top, float width, float height, float x, float y, float xscale, float yscale, double rot, int c1, int c2, int c3, int c4, double a1, double a2, double a3, double a4);
 
-void draw_sprite_tiled(int spr,int subimg,double x,double y);
-void draw_sprite_tiled_ext(int spr,int subimg,double x,double y,double xscale,double yscale,int color,double alpha);
+void draw_sprite_tiled(int spr, int subimg, float x, float y);
+void draw_sprite_tiled_ext(int spr, int subimg, float x, float y, float xscale, float yscale, int color, double alpha);
 
 bool sprite_exists(int sprite);
