+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGLES/GLESbackground.cpp
@@ -26,7 +26,7 @@
 extern int room_width, room_height;
 namespace enigma{extern unsigned bound_texture;}
 
-int draw_background(int back, double x, double y)
+int draw_background(int back, float x, float y)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -60,7 +60,7 @@ int draw_background(int back, double x, double y)
   return 0;
 }
 
-int draw_background_stretched(int back, double x, double y, double w, double h)
+int draw_background_stretched(int back, float x, float y, float w, float h)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -94,7 +94,7 @@ int draw_background_stretched(int back, double x, double y, double w, double h)
   return 0;
 }
 
-int draw_background_part(int back,double left,double top,double width,double height,double x,double y)
+int draw_background_part(int back, float left, float top, float width, float height, float x, float y)
 {
     enigma::background *bck2d = enigma::backgroundstructarray[back];
     if (!bck2d)
@@ -129,7 +129,7 @@ int draw_background_part(int back,double left,double top,double width,double hei
     return 0;
 }
 
-int draw_background_tiled(int back,double x,double y)
+int draw_background_tiled(int back, float x, float y)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -172,7 +172,7 @@ int draw_background_tiled(int back,double x,double y)
   return 0;
 }
 
-int draw_background_tiled_area(int back,double x,double y,double x1,double y1,double x2,double y2)
+int draw_background_tiled_area(int back, float x, float y, float x1, float y1, float x2, float y2)
 {
     enigma::background *bck2d = enigma::backgroundstructarray[back];
     if (!bck2d)
@@ -235,7 +235,7 @@ int draw_background_tiled_area(int back,double x,double y,double x1,double y1,do
   return 0;
 }
 
-int draw_background_ext(int back,double x,double y,double xscale,double yscale,double rot,int color,double alpha)
+int draw_background_ext(int back, float x, float y, float xscale, float yscale, double rot, int color, double alpha)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -275,7 +275,7 @@ int draw_background_ext(int back,double x,double y,double xscale,double yscale,d
   return 0;
 }
 
-int draw_background_stretched_ext(int back,double x,double y,double w,double h,int color,double alpha)
+int draw_background_stretched_ext(int back, float x, float y, float w, float h, int color, double alpha)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -309,7 +309,7 @@ int draw_background_stretched_ext(int back,double x,double y,double w,double h,i
   return 0;
 }
 
-int draw_background_part_ext(int back,double left,double top,double width,double height,double x,double y,double xscale,double yscale,int color,double alpha)
+int draw_background_part_ext(int back, float left, float top, float width, float height, float x, float y, float xscale, float yscale, int color, double alpha)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -346,7 +346,7 @@ int draw_background_part_ext(int back,double left,double top,double width,double
   return 0;
 }
 
-int draw_background_tiled_ext(int back,double x,double y,double xscale,double yscale,int color,double alpha)
+int draw_background_tiled_ext(int back, float x, float y, float xscale, float yscale, int color, double alpha)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -388,7 +388,7 @@ int draw_background_tiled_ext(int back,double x,double y,double xscale,double ys
   return 0;
 }
 
-int draw_background_tiled_area_ext(int back,double x,double y,double x1,double y1,double x2,double y2, double xscale, double yscale, int color, double alpha)
+int draw_background_tiled_area_ext(int back, float x, float y, float x1, float y1, float x2, float y2, float xscale, float yscale, int color, double alpha)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
@@ -450,7 +450,7 @@ int draw_background_tiled_area_ext(int back,double x,double y,double x1,double y
   return 0;
 }
 
-int draw_background_general(int back,double left,double top,double width,double height,double x,double y,double xscale,double yscale,double rot,int c1,int c2,int c3,int c4,double a1,double a2,double a3,double a4)
+int draw_background_general(int back, float left, float top, float width, float height, float x, float y, float xscale, float yscale, double rot, int c1, int c2, int c3, int c4, double a1, double a2, double a3, double a4)
 {
   enigma::background *bck2d = enigma::backgroundstructarray[back];
   if (!bck2d)
